create ec.go
