+++ openssl/evp.go
@@ -13,16 +13,6 @@ import (
 	"unsafe"
 )
 
-var (
-	paramAlgHMAC = C.CString("HMAC")
-	paramDigest  = C.CString("digest")
-	paramPubKey  = C.CString("pub")
-	paramPrivKey = C.CString("priv")
-	paramGroup   = C.CString("group")
-	paramECPubX  = C.CString("qx")
-	paramECPubY  = C.CString("qy")
-)
-
 // hashToMD converts a hash.Hash implementation from this package to a GO_EVP_MD_PTR.
 func hashToMD(h hash.Hash) C.GO_EVP_MD_PTR {
 	switch h.(type) {
@@ -402,35 +392,4 @@ func newEvpFromParams(id C.int, selection C.int, params C.GO_OSSL_PARAM_PTR) (C.
 		return nil, newOpenSSLError("EVP_PKEY_fromdata")
 	}
 	return pkey, nil
-}
-
-// encodeEcPoint encodes pt.
-func encodeEcPoint(group C.GO_EC_GROUP_PTR, pt C.GO_EC_POINT_PTR) ([]byte, error) {
-	// Get encoded point size.
-	n := C.go_openssl_EC_POINT_point2oct(group, pt, C.GO_POINT_CONVERSION_UNCOMPRESSED, nil, 0, nil)
-	if n == 0 {
-		return nil, newOpenSSLError("EC_POINT_point2oct")
-	}
-	// Encode point into bytes.
-	bytes := make([]byte, n)
-	n = C.go_openssl_EC_POINT_point2oct(group, pt, C.GO_POINT_CONVERSION_UNCOMPRESSED, base(bytes), n, nil)
-	if n == 0 {
-		return nil, newOpenSSLError("EC_POINT_point2oct")
-	}
-	return bytes, nil
-}
-
-// generateAndEncodePublicKey calls newPubKeyPointFn to generate a public key point and then encodes it.
-func generateAndEncodePublicKey(nid C.int, newPubKeyPointFn func(group C.GO_EC_GROUP_PTR) (C.GO_EC_POINT_PTR, error)) ([]byte, error) {
-	group := C.go_openssl_EC_GROUP_new_by_curve_name(nid)
-	if group == nil {
-		return nil, newOpenSSLError("EC_GROUP_new_by_curve_name")
-	}
-	defer C.go_openssl_EC_GROUP_free(group)
-	pt, err := newPubKeyPointFn(group)
-	if err != nil {
-		return nil, err
-	}
-	defer C.go_openssl_EC_POINT_free(pt)
-	return encodeEcPoint(group, pt)
 }
